django_with_vue/vue_cor_jwt/server/requirements.txt
asgiref==3.4.1
Django==3.2.6
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
PyJWT==2.3.0
python-decouple==3.5
pytz==2021.1
sqlparse==0.4.1
typing-extensions==4.0.0
